requirements.txt.txt
pathway
requests
datetime
openai
together ai
python-dotenv
feedparser
beautifulsoup4